:ambulance: 한글로 tag 하였을 때 오류 발생하여 개선
